Add session to get error messages from user service jobs controller to schedule controller
Copy the alert component styling to the schedule index to use the
array of errors in the schedule page

add validations and custom validation to make sure the time
inputs are proper.

The alert component expects an object but since I am using a
session to store the messages in an array I could not use the
component itself but could use the styling.