Change section tags to div to fix failed submissions due to submit button being outside of the form.
I could not find where the tags were misaligned or out of order but I did find
that the seciton tags do not belong in forms and swapped them. The submit button
is now inside of the form. Tested in chrome, firefox, and safari.
the commit message for your changes. Lines
starting # with '#' will be ignored, and an empty message aborts the commit.